backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sentence-transformers==3.1.1
torch==2.4.1
transformers==4.45.2
huggingface-hub==0.25.1
chromadb==0.5.5
hnswlib==0.8.0
groq==0.9.0
pydantic==2.9.2
aiofiles==24.1.0
numpy==1.26.4
